// src/api/v1/shared/Prefecture_pb.ts
export enum Prefecture {
  /**
   * @generated from enum value: Unspecified = 0;
   */
  Unspecified = 0,

  /**
   * @generated from enum value: Hokkaido = 1;
   */
  Hokkaido = 1,

  /**
   * @generated from enum value: Aomori = 2;
   */
  Aomori = 2,

  /**
   * @generated from enum value: Iwate = 3;
   */
  Iwate = 3,

  /**
   * @generated from enum value: Miyagi = 4;
   */
  Miyagi = 4,

  /**
   * @generated from enum value: Akita = 5;
   */
  Akita = 5,

  /**
   * @generated from enum value: Yamagata = 6;
   */
  Yamagata = 6,

  /**
   * @generated from enum value: Fukushima = 7;
   */
  Fukushima = 7,

  /**
   * @generated from enum value: Ibaraki = 8;
   */
  Ibaraki = 8,

  /**
   * @generated from enum value: Tochigi = 9;
   */
  Tochigi = 9,

  /**
   * @generated from enum value: Gunma = 10;
   */
  Gunma = 10,

  /**
   * @generated from enum value: Saitama = 11;
   */
  Saitama = 11,

  /**
   * @generated from enum value: Chiba = 12;
   */
  Chiba = 12,

  /**
   * @generated from enum value: Tokyo = 13;
   */
  Tokyo = 13,

  /**
   * @generated from enum value: Kanagawa = 14;
   */
  Kanagawa = 14,

  /**
   * @generated from enum value: Niigata = 15;
   */
  Niigata = 15,

  /**
   * @generated from enum value: Toyama = 16;
   */
  Toyama = 16,

  /**
   * @generated from enum value: Ishikawa = 17;
   */
  Ishikawa = 17,

  /**
   * @generated from enum value: Fukui = 18;
   */
  Fukui = 18,

  /**
   * @generated from enum value: Yamanashi = 19;
   */
  Yamanashi = 19,

  /**
   * @generated from enum value: Nagano = 20;
   */
  Nagano = 20,

  /**
   * @generated from enum value: Gifu = 21;
   */
  Gifu = 21,

  /**
   * @generated from enum value: Shizuoka = 22;
   */
  Shizuoka = 22,

  /**
   * @generated from enum value: Aichi = 23;
   */
  Aichi = 23,

  /**
   * @generated from enum value: Mie = 24;
   */
  Mie = 24,

  /**
   * @generated from enum value: Shiga = 25;
   */
  Shiga = 25,

  /**
   * @generated from enum value: Kyoto = 26;
   */
  Kyoto = 26,

  /**
   * @generated from enum value: Osaka = 27;
   */
  Osaka = 27,

  /**
   * @generated from enum value: Hyogo = 28;
   */
  Hyogo = 28,

  /**
   * @generated from enum value: Nara = 29;
   */
  Nara = 29,

  /**
   * @generated from enum value: Wakayama = 30;
   */
  Wakayama = 30,

  /**
   * @generated from enum value: Tottori = 31;
   */
  Tottori = 31,

  /**
   * @generated from enum value: Shimane = 32;
   */
  Shimane = 32,

  /**
   * @generated from enum value: Okayama = 33;
   */
  Okayama = 33,

  /**
   * @generated from enum value: Hiroshima = 34;
   */
  Hiroshima = 34,

  /**
   * @generated from enum value: Yamaguchi = 35;
   */
  Yamaguchi = 35,

  /**
   * @generated from enum value: Tokushima = 36;
   */
  Tokushima = 36,

  /**
   * @generated from enum value: Kagawa = 37;
   */
  Kagawa = 37,

  /**
   * @generated from enum value: Ehime = 38;
   */
  Ehime = 38,

  /**
   * @generated from enum value: Kochi = 39;
   */
  Kochi = 39,

  /**
   * @generated from enum value: Fukuoka = 40;
   */
  Fukuoka = 40,

  /**
   * @generated from enum value: Saga = 41;
   */
  Saga = 41,

  /**
   * @generated from enum value: Nagasaki = 42;
   */
  Nagasaki = 42,

  /**
   * @generated from enum value: Kumamoto = 43;
   */
  Kumamoto = 43,

  /**
   * @generated from enum value: Oita = 44;
   */
  Oita = 44,

  /**
   * @generated from enum value: Miyazaki = 45;
   */
  Miyazaki = 45,

  /**
   * @generated from enum value: Kagoshima = 46;
   */
  Kagoshima = 46,

  /**
   * @generated from enum value: Okinawa = 47;
   */
  Okinawa = 47,
}